+++ packages/stylelint-config-copilot-base/index.js
@@ -9,27 +9,6 @@ module.exports = {
     /* These rules should be configured on an editor level (e.g. with editorconfig) */
     /* start */
 
-    // Specify indentation.
-    'indentation': null,
-
-    // Disallow end-of-line whitespace.
-    'no-eol-whitespace': true,
-
-    // Disallow missing end-of-source newlines.
-    'no-missing-end-of-source-newline': true,
-
-    // Specify unix or windows linebreaks.
-    'linebreaks': null,
-    /* end */
-
-    // Require or disallow the Unicode Byte Order Mark.
-    'unicode-bom': null,
-
-    'max-empty-lines': 5,
-
-    // Limit the length of a line.
-    'max-line-length': null,
-
     // Disallow irregular whitespaces.
     'no-irregular-whitespace': null,
 
@@ -55,11 +34,6 @@ module.exports = {
 
     // Disallow empty sources.
     'no-empty-source': true,
-
-    'no-empty-first-line': null,
-
-    // Disallow extra semicolons.
-    'no-extra-semicolons': true,
 
     // Disallow duplicate selectors within a stylesheet.
     'no-duplicate-selectors': true,
@@ -87,9 +61,6 @@ module.exports = {
 
     // Require or disallow alpha channel for hex colors.
     'color-hex-alpha': null,
-
-    // Specify lowercase or uppercase for hex colors.
-    'color-hex-case': 'upper',
 
     // Specify short or long notation for hex colors.
     // Hex colours must always use shorthand where possible.
@@ -125,9 +96,13 @@ module.exports = {
     // Specify number or angle notation for degree hues
     'hue-degree-notation': null,
 
+    // Specify number or percentage notation for lightness.
+    'lightness-notation': null,
+
     /* ------------------------------------*\
             #ANNOTATION
     \*------------------------------------ */
+    // Disallow unknown annotations.
     'annotation-no-unknown': true,
 
     /* ------------------------------------*\
@@ -163,22 +138,6 @@ module.exports = {
     // Disallow an unspaced operator within calc functions.
     'function-calc-no-unspaced-operator': true,
 
-    // Require a newline or disallow whitespace after the commas of functions.
-    // There must always be a newline after the commas in multi-line functions.
-    'function-comma-newline-after': 'never-multi-line',
-
-    // Require a newline or disallow whitespace before the commas of functions.
-    // There must never be a whitespace before the commas in multi-line functions.
-    'function-comma-newline-before': 'never-multi-line',
-
-    // Require a single space or disallow whitespace after the commas of functions.
-    // There must always be a single space after the commas in single-line functions
-    'function-comma-space-after': 'always-single-line',
-
-    // Require a single space or disallow whitespace before the commas of functions.
-    // There must never be whitepace before the commas in single-line functions.
-    'function-comma-space-before': 'never',
-
     // Disallow direction values in linear-gradient() calls that are not valid according to the standard syntax.
     'function-linear-gradient-no-nonstandard-direction': [
       true,
@@ -187,19 +146,8 @@ module.exports = {
       },
     ],
 
-    // Limit the number of adjacent empty lines within functions.
-    'function-max-empty-lines': 0,
-
     // Specify lowercase or uppercase for function names.
     'function-name-case': 'lower',
-
-    // Require a newline or disallow whitespace on the inside of the parentheses of functions.
-    // There must always be a newline if the function call is multiline.
-    'function-parentheses-newline-inside': 'never-multi-line',
-
-    // Require a single space or disallow whitespace on the inside of the parentheses of functions.
-    // There must never be whitespace on the inside the parentheses.
-    'function-parentheses-space-inside': 'never',
 
     // Require or disallow data URIs for urls.
     // "function-url-data-uris": "always",
@@ -210,10 +158,6 @@ module.exports = {
 
     // Specify a allowed-list of only allowed functions.
     // "function-allowed-list": false,
-
-    // Require a single space or disallow whitespace after functions.
-    // There must always be a single space after the function.
-    'function-whitespace-after': 'always',
 
     // Disallow scheme-relative urls.
     'function-url-no-scheme-relative': null,
@@ -235,13 +179,6 @@ module.exports = {
     /* ------------------------------------*\
             #NUMBER
     \*------------------------------------ */
-    // Require or disallow a leading zero for fractional numbers less than 1.
-    // There must always be a leading zero.
-    'number-leading-zero': 'always',
-
-    // Disallow trailing zeros within numbers.
-    'number-no-trailing-zeros': true,
-
     // Disallow units for zero lengths.
     // "number-zero-length-no-unit": true,
 
@@ -253,10 +190,6 @@ module.exports = {
     \*------------------------------------ */
     // Disallow (unescaped) newlines in strings.
     'string-no-newline': true,
-
-    // Specify single or double quotes around strings
-    // Strings must always be wrapped with single quotes.
-    'string-quotes': 'double',
 
     /* ------------------------------------*\
             #TIME
@@ -274,14 +207,6 @@ module.exports = {
 
     // Specify a allowed-list of allowed units.
     'unit-allowed-list': null,
-
-    // Specify lowercase or uppercase for units.
-    'unit-case': [
-      'lower',
-      {
-        'severity': 'error',
-      },
-    ],
 
     // Specify a allowed-list of allowed units.
     // "unit-allowed-list": false,
@@ -311,42 +236,12 @@ module.exports = {
     //   },
     // ],
 
-    // Require a newline or disallow whitespace after the commas of value lists.
-    // There must always be a newline after the commas in multi-line value lists.
-    'value-list-comma-newline-after': 'always-multi-line',
-
-    // Require a newline or disallow whitespace before the commas of value lists.
-    // There must always be whitespace before the commas in multi-line value lists.
-    'value-list-comma-newline-before': 'never-multi-line',
-
-    // Require a single space or disallow whitespace after the commas of value lists.
-    // There must always be a single space after the commas in single-line value lists.
-    'value-list-comma-space-after': 'always-single-line',
-
-    // Require a single space or disallow whitespace before the commas of value lists.
-    // There must never be whitespace before the commas in single-line value lists.
-    'value-list-comma-space-before': 'never',
-
-    // Limit the number of adjacent empty lines within value lists.
-    'value-list-max-empty-lines': 0,
-
     /* ------------------------------------*\
             #SHORTHAND
     \*------------------------------------ */
     // # Shorthand property
     // Disallow redundant values in shorthand properties.
     'shorthand-property-no-redundant-values': true,
-
-    /* ------------------------------------*\
-            #PROPERTY
-    \*------------------------------------ */
-    // Specify lowercase or uppercase for properties.
-    'property-case': [
-      'lower',
-      {
-        'severity': 'error',
-      },
-    ],
 
     // Disallow vendor prefixes for properties.
     'property-no-vendor-prefix': true,
@@ -368,22 +263,6 @@ module.exports = {
     /* ------------------------------------*\
             #DECLARATION
     \*------------------------------------ */
-    // Require a single space or disallow whitespace after the bang of declarations.
-    // There must never be whitespace after the bang.
-    'declaration-bang-space-after': 'never',
-
-    // Require a single space or disallow whitespace before the bang of declarations.
-    // There must always be a single space before the bang.
-    'declaration-bang-space-before': 'always',
-
-    // Require a single space or disallow whitespace after the colon of declarations.
-    // There must always be a single space after the colon if the declaration"s value is single-line.
-    'declaration-colon-space-after': 'always-single-line',
-
-    // Require a single space or disallow whitespace before the colon of declarations.
-    // There must never be whitespace before the colon
-    'declaration-colon-space-before': 'never',
-
     // Disallow !important within declarations.
     'declaration-no-important': true,
 
@@ -391,7 +270,6 @@ module.exports = {
     'declaration-block-no-redundant-longhand-properties': null,
 
     // Require a newline or disallow whitespace after the colon of declarations.
-    'declaration-colon-newline-after': null,
     'declaration-empty-line-before': null,
     'declaration-property-unit-disallowed-list': null,
     'declaration-property-unit-allowed-list': null,
@@ -402,6 +280,10 @@ module.exports = {
         ],
       },
     ],
+
+    // Disallow unknown values for properties within declarations.
+    'declaration-property-value-no-unknown': null,
+
     // Specify a list of allowed property and value pairs within declarations.
     'declaration-property-value-allowed-list': null,
 
@@ -433,57 +315,12 @@ module.exports = {
     // Disallow duplicate custom properties within declaration blocks.
     'declaration-block-no-duplicate-custom-properties': true,
 
-    // Require a newline or disallow whitespace after the semicolons of declaration blocks.
-    // There must always be a newline after the semicolon.
-    'declaration-block-semicolon-newline-after': 'always',
-
-    // Require a newline or disallow whitespace before the semicolons of declaration blocks.
-    // There must never be whitespace before the semicolons in multi-line rules.
-    'declaration-block-semicolon-newline-before': 'never-multi-line',
-
-    // Require a single space or disallow whitespace after the semicolons of declaration blocks.
-    // There must always be a single space after the semicolon in single-line declaration blocks.
-    'declaration-block-semicolon-space-after': 'always-single-line',
-
-    // Require a single space or disallow whitespace before the semicolons of declaration blocks.
-    // There must never be whitespace before the semicolons.
-    'declaration-block-semicolon-space-before': 'never',
-
     // Limit the number of declaration within single line declaration blocks.
     'declaration-block-single-line-max-declarations': 1,
-
-    // Require or disallow a trailing semicolon within declaration blocks.
-    'declaration-block-trailing-semicolon': 'always',
 
     /* ------------------------------------*\
             #BLOCK
     \*------------------------------------ */
-    // Require or disallow an empty line before the closing brace of blocks.
-    'block-closing-brace-empty-line-before': null,
-
-    // Require a newline or disallow whitespace after the closing brace of blocks.
-    // There must always be a newline after the closing brace.
-    'block-closing-brace-newline-after': [
-      'always',
-      {
-        'ignoreAtRules': [
-          'if',
-          'else',
-        ],
-      },
-    ],
-
-    // Require a newline or disallow whitespace before the closing brace of blocks.
-    // There must always be a newline before the closing brace in multi-line blocks.
-    'block-closing-brace-newline-before': 'always-multi-line',
-
-    // Require a single space or disallow whitespace after the closing brace of blocks.
-    // There must always be a single space after the closing brace in single-line blocks.
-    'block-closing-brace-space-after': 'always-single-line',
-
-    // Require a single space or disallow whitespace before the closing brace of blocks.
-    'block-closing-brace-space-before': null,
-
     // Require a single space or disallow whitespace before the closing brace of blocks.
     // There must always be a single space before the closing brace in single-line blocks.
     // "block-closing-brace-space-before": "always-single-line",
@@ -491,30 +328,11 @@ module.exports = {
     // Disallow empty blocks.
     'block-no-empty': true,
 
-    // Require a newline after the opening brace of blocks.
-    // There must always be a newline after the opening brace.
-    'block-opening-brace-newline-after': null,
-
-    // Require a newline or disallow whitespace before the opening brace of blocks.
-    // (this rule is totally bugged out and does not work properly)
-    'block-opening-brace-newline-before': 'never-single-line',
-
-    // Require a single space or disallow whitespace after the opening brace of blocks.
-    // There must always be a single space after the opening brace in single-line blocks.
-    'block-opening-brace-space-after': 'always-single-line',
-
-    // Require a single space or disallow whitespace before the opening brace of blocks.
-    // There must always be a single space before the opening brace.
-    'block-opening-brace-space-before': 'always',
-
     /* ------------------------------------*\
             #SELECTOR
     \*------------------------------------ */
     // Specify a list of disallowed attribute names.
     'selector-attribute-name-disallowed-list': null,
-
-    // Require a single space or disallow whitespace on the inside of the brackets within attribute selectors.
-    'selector-attribute-brackets-space-inside': 'never',
 
     // Specify a disallowed-list of disallowed attribute operators.
     // "selector-attribute-operator-disallowed-list": false,
@@ -522,28 +340,11 @@ module.exports = {
     // Specify a allowed-list of allowed attribute operators.
     // "selector-attribute-operator-allowed-list": false,
 
-    // Require a single space or disallow whitespace after operators within attribute selectors.
-    'selector-attribute-operator-space-after': 'never',
-
-    // Require a single space or disallow whitespace before operators within attribute selectors.
-    'selector-attribute-operator-space-before': 'never',
-
     // Require or disallow quotes for attribute values.
     'selector-attribute-quotes': 'always',
 
     // Specify a pattern for class selectors.
     // "selector-class-pattern": /^.([a-z](-[a-z])?)+(__([a-z](-[a-z])?)+)?(--([a-z](-[a-z])?)+)?$/,
-
-    // Require a single space or disallow whitespace after the combinators of selectors.
-    // There must always be a single space after the combinators.
-    'selector-combinator-space-after': 'always',
-
-    // Require a single space or disallow whitespace before the combinators of selectors.
-    // There must always be a single space before the combinators.
-    'selector-combinator-space-before': 'always',
-
-    // Disallow non-space characters for descendant combinators of selectors.
-    'selector-descendant-combinator-no-non-space': true,
 
     // Limit the number of attribute selectors in a selector.
     'selector-max-attribute': 2,
@@ -566,9 +367,6 @@ module.exports = {
         'severity': 'warning',
       },
     ],
-
-    // Limit the number of adjacent empty lines within selectors.
-    'selector-max-empty-lines': 0,
 
     // Limit the number of id selectors in a selector.
     'selector-max-id': 0,
@@ -602,16 +400,8 @@ module.exports = {
     // Disallow unknown pseudo-class selectors.
     'selector-pseudo-class-no-unknown': true,
 
-    // Specify lowercase or uppercase for pseudo-class selectors.
-    'selector-pseudo-class-case': 'lower',
-
-    // Require a single space or disallow whitespace on the inside of the parentheses within pseudo-class selectors.
-    'selector-pseudo-class-parentheses-space-inside': 'never',
-
     // Disallow unknown pseudo-element selectors.
     'selector-pseudo-element-no-unknown': true,
-    // Specify lowercase or uppercase for pseudo-element selectors.
-    'selector-pseudo-element-case': 'lower',
 
     // Specify single or double colon notation for applicable pseudo-elements.
     // Applicable pseudo-elements must always use the double colon notation.
@@ -665,24 +455,8 @@ module.exports = {
     // Specify simple or complex notation for :not() pseudo-class selectors.
     'selector-not-notation': null,
 
-    /* ------------------------------------*\
-            #SELECTOR-LIST
-    \*------------------------------------ */
-    // Require a newline or disallow whitespace after the commas of selector lists.
-    // There must always be a newline after the commas in multi-line selector lists.
-    'selector-list-comma-newline-after': 'always-multi-line',
-
-    // Require a newline or disallow whitespace before the commas of selector lists.
-    // There must never be whitespace before the commas in multi-line selector lists.
-    'selector-list-comma-newline-before': 'never-multi-line',
-
-    // Require a single space or disallow whitespace after the commas of selector lists.
-    // There must always be a single space after the commas in single-line selector lists.
-    'selector-list-comma-space-after': 'always-single-line',
-
-    // Require a single space or disallow whitespace before the commas of selector lists.
-    // There must never be whitespace before the commas.
-    'selector-list-comma-space-before': 'never',
+    // Disallow unmatchable An+B selectors.
+    'selector-anb-no-unmatchable': true,
 
     /* ------------------------------------*\
             #RULE
@@ -707,20 +481,6 @@ module.exports = {
     /* ------------------------------------*\
             #MEDIA-FEATURE
     \*------------------------------------ */
-    // Require a single space or disallow whitespace after the colon in media features.
-    // There must always be a single space after the colon.
-    'media-feature-colon-space-after': 'always',
-
-    // Require a single space or disallow whitespace before the colon in media features.
-    // There must never be whitespace before the colon.
-    'media-feature-colon-space-before': 'never',
-
-    // Specify lowercase or uppercase for media feature names.
-    'media-feature-name-case': 'lower',
-
-    // Require a single space or disallow whitespace on the inside of the parentheses within media features.
-    'media-feature-parentheses-space-inside': null,
-
     // Disallow vendor prefixes for media feature names.
     'media-feature-name-no-vendor-prefix': null,
 
@@ -736,33 +496,17 @@ module.exports = {
     // Disallow unknown media feature names.
     'media-feature-name-no-unknown': true,
 
-    // Require a single space or disallow whitespace after the range operator in media features (Autofixable).
-    'media-feature-range-operator-space-after': 'always',
-
-    // Require a single space or disallow whitespace before the range operator in media features (Autofixable).
-    'media-feature-range-operator-space-before': 'always',
-
     // Specify context or prefix notation for media feature ranges.
     'media-feature-range-notation': null,
 
-    /* ------------------------------------*\
-            #MEDIA-QUERY-LIST
-    \*------------------------------------ */
-    // Require a newline or disallow whitespace after the commas of media query lists.
-    // There must always be a newline after the commas in multi-line media query lists.
-    'media-query-list-comma-newline-after': 'never-multi-line',
+    // Specify a list of allowed name and unit pairs within media features.
+    'media-feature-name-unit-allowed-list': null,
 
-    // Require a newline or disallow whitespace before the commas of media query lists.
-    // There must never be a white before the commas in multi-line media query lists.
-    'media-query-list-comma-newline-before': 'never-multi-line',
+    // Disallow unknown values for media features.
+    'media-feature-name-value-no-unknown': true,
 
-    // Require a single space or disallow whitespace after the commas of media query lists.
-    // There must always be a single space after the commas in single-line media query lists.
-    'media-query-list-comma-space-after': 'always-single-line',
-
-    // Require a single space or disallow whitespace before the commas of media query lists.
-    // There must never be whitepace before the commas.
-    'media-query-list-comma-space-before': 'never',
+    // Disallow invalid media queries.
+    'media-query-no-invalid': true,
 
     /* ------------------------------------*\
             #AT-RULE
@@ -805,23 +549,8 @@ module.exports = {
     // Specify a allowed-list of allowed at-rules.
     'at-rule-allowed-list': null,
 
-    // Specify lowercase or uppercase for at-rules names.
-    'at-rule-name-case': 'lower',
-
     // Disallow unknown at-rules.
     'at-rule-no-unknown': true,
-
-    // Require a newline after at-rule names.
-    'at-rule-name-newline-after': null,
-
-    // Require a single space after at-rule names.
-    'at-rule-name-space-after': 'always',
-
-    // Require a newline after the semicolon of at-rules.
-    'at-rule-semicolon-newline-after': 'always',
-
-    // Require a single space or disallow whitespace before the semicolons at-rules.
-    'at-rule-semicolon-space-before': 'never',
 
     /* ------------------------------------*\
             #ANIMATIONS / KEYFRAMES
@@ -881,6 +610,12 @@ module.exports = {
 
     // Disallow missing var function for custom properties.
     'custom-property-no-missing-var-function': true,
+
+    // Disallow unknown custom properties.
+    // @TODO:
+    // evaluate 'This rule considers custom properties defined within the same source to be known.'
+    // https://stylelint.io/user-guide/rules/no-unknown-custom-properties/
+    'no-unknown-custom-properties': null,
 
     /* ------------------------------------*\
             #GRID