+++ packages/stylelint-config-copilot-scss/index.js
@@ -11,9 +11,24 @@ module.exports = {
     // Adjust default rules
     'function-no-unknown': null,
 
+    // Disallow unknown annotations.
+    'annotation-no-unknown': [
+      true,
+      {
+        'ignoreAnnotations': [
+          'default',
+          'global',
+          'optional',
+          '/^=/',
+        ],
+      },
+    ],
+
     /* ------------------------------------*\
       #SCSS-RULES
     \*------------------------------------ */
+    // Disallow nesting a single block if it could be merged with its parent block.
+    'scss/block-no-redundant-nesting': null,
 
     /* ------------------------------------*\
       #AT-IF
@@ -78,15 +93,15 @@ module.exports = {
     \*------------------------------------ */
 
     // Disallow leading underscore in partial names in @import.
-    'scss/at-import-no-partial-leading-underscore': null,
+    'scss/load-no-partial-leading-underscore': null,
 
     // Specify disallowed-list of disallowed file extensions for partial names in @import commands.
-    'scss/at-import-partial-extension-blacklist': [
+    'scss/at-import-partial-extension-disallowed-list': [
       'scss',
     ],
 
-    // Specify whitelist of allowed file extensions for partial names in @import commands.
-    'scss/at-import-partial-extension-whitelist': null,
+    // Specify list of allowed file extensions for partial names in @import commands.
+    'scss/at-import-partial-extension-allowed-list': null,
 
     // Require or disallow extension in @import commands.
     'scss/at-import-partial-extension': null,
@@ -121,7 +136,16 @@ module.exports = {
 
     // Disallow unknown at-rules. Should be used instead of stylelint's at-rule-no-unknown.
     'at-rule-no-unknown': null,
-    'scss/at-rule-no-unknown': true,
+
+    // Allow container queries
+    'scss/at-rule-no-unknown': [
+      true,
+      {
+        ignoreAtRules: [
+          'container',
+        ],
+      },
+    ],
 
     'at-rule-disallowed-list': [
       'debug',
@@ -133,14 +157,26 @@ module.exports = {
     // Disallow usage of @use without a namespace.
     'scss/at-use-no-unnamespaced': null,
 
+    // Disallow redundant namespace aliases.
+    'scss/at-use-no-redundant-alias': true,
+
+    // Disallow redundant @at-root rule.
+    'scss/at-root-no-redundant': true,
+
     /* ------------------------------------*\
       #FUNCTIONS
     \*------------------------------------ */
     // This rule is basically a wrapper around the mentioned core rule. You must disable Stylelint's core rule to make this rule work
-    'scss/function-no-unknown': true,
+    'scss/function-no-unknown': null,
+
+    // Specify a list of disallowed functions.
+    'scss/function-disallowed-list': null,
 
     // Disallows the use of global function names, as these global functions are now located inside built-in Sass modules.
     'scss/no-global-function-names': null,
+
+    // Disallow unused private members.
+    'scss/no-unused-private-members': null,
 
     // Encourage the use of the scale-color over sass color-functions (darken, lighten, etc.).
     'scss/function-color-relative': null,
@@ -150,6 +186,9 @@ module.exports = {
 
     // Disallow unquoted strings inside the unquote function.
     'scss/function-unquote-no-unquoted-strings-inside': null,
+
+    // Disallow interpolation in calc(), clamp(), min(), and max() functions.
+    'scss/function-calculation-no-interpolation': null,
 
     // Require for variables to be put first in a block (a rule or in root)
     'scss/dollar-variable-first-in-block': [
@@ -315,6 +354,10 @@ module.exports = {
     /* ------------------------------------*\
       #GENERAL
     \*------------------------------------ */
+
+    // Disallow unknown properties. Should be used instead of Stylelint's property-no-unknown.
+    'property-no-unknown': null,
+    'scss/property-no-unknown': true,
 
     // Disallow dollar variables within a stylesheet.
     'scss/no-dollar-variables': null,