+++ packages/stylelint-config-copilot-order/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-rc.4] - 22.04.204
+- chore:
+  - upgrade to Stylelint 16 (see [migration guide](https://github.com/stylelint/stylelint/blob/main/docs/migration-guide/to-16.md))
+- rules:
+  - group `gaps` below flexbox and grid properties
+
 ## [1.0.0-rc.3] - 31.01.2023
 - update deps