+++ packages/stylelint-config-copilot-plugins/CHANGELOG.md
@@ -4,6 +4,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-rc.4] - 22.04.204
+- breaking changes:
+  - stylelint-itcss: remove plugin as it is not supported for Stylelint versions > 14
+    - alternatively use [Stylelint's overrides](https://stylelint.io/user-guide/configure/#overrides), e.g. to disable `declaration-no-important` in utilities layer
+      ```
+      overrides: [
+        {
+            files: [
+                '<path/to/utility-layer>/**/*.scss',
+            ],
+            rules: {
+                'declaration-no-important': null,
+            },
+        }
+      ]
+      ```
+- chore:
+  - upgrade to Stylelint 16 (see [migration guide](https://github.com/stylelint/stylelint/blob/main/docs/migration-guide/to-16.md))
+
+
 ## [1.0.0-rc.3] - 31.01.2023
 - update deps